requirements.txt
# PyTorch 深度学习框架
torch>=1.12.0
torchvision>=0.13.0

# 数据处理
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0

# 配置管理
pyyaml>=6.0

# 进度条和日志
tqdm>=4.62.0

# 可视化
matplotlib>=3.5.0
tensorboard>=2.11.0

# 数据分析
pandas>=1.3.0

# YOLO (可选，用于emotion_recognition项目)
ultralytics>=8.0.0
